LayoutTests/security/contentSecurityPolicy/block-javascripturl-non-inline-csp-expected.txt

CONSOLE MESSAGE: Script is executing and is not blocked by CSP.
CONSOLE MESSAGE: Refused to execute a script because its hash or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
Test output should contain a CSP violation.
